Create ``_MessageStateHandler`` and move all private methods
